Remove workaround for Hawkular-Metrics before v0.8.2
Hawkular-Metrics before version 0.8.2 did not export an object's type to
the tags. Now that all clusters are using more recent versions this
workaround can be removed.

This removal has the intended side-effect of fixing an issue whereby
empty metrics produced an error messages due to there not being a "tags"
member (“KeyError: 'tags'”).