Show storage class for pending and lost volume claims
Amend the "check_openshift_pvc_phase" script to show the requested
storage class for pending and lost claims. This information can be
useful in resolving issues and save the operator from having to look it
up on the cluster.

Also change the separator between individual PVCs from a space to
a comma.