Add age limit to pending PVCs before going critical
When a volume claim is critical for an extended period of time it's
usually indicative of there not being enough volumes or dynamic creation
not working. Amend the "check_openshift_pvc_phase" script to fail with
a critical state when at least one claim is over an hour old (duration
configurable).