Replace lambdas with direct runtime macro access
The arguments for the "openshift_node_resources" check command used
"skip_if" with a lambda. The intention was to only set arguments when
they expanded to a string at least one character in length.

Most deployments configure the variables, i.e.
"openshift_node_cpu_limits_warn", to numbers. Those deployments never
saw limit arguments to "openshift_node_resources" applied.

The "macro" function does not automatically convert values to strings.
Instead the Icinga2 script language returns them in their original type
if and only if they're the only expanded macro ("macro("$foo$=$bar$")"
produces a string with "foo" and "bar" expanded, "macro("$num$")"'s
return type depends on the expanded variable).

The global "len" function is not recommended and the Icinga2
documentation[1] advises to use type-specific functions instead (i.e.
String#len).

The implementation of the global "len" function silently returns 0 for
unrecognized types such as Number. In this case the lambda given to
"skip_if" would call "len" on a Number, always getting a length of zero
and thus skipping arguments.

One option would've been to use the "string" function, i.e.
"len(string(macro("$…$"))) > 0". Another is to do away with the lambdas
altogether and to rely on the implicit conversion to Boolean[3] instead.
The latter option is implemented with this change.

[1]
https://www.icinga.com/docs/icinga2/latest/doc/18-library-reference/#len

[2]
https://github.com/Icinga/icinga2/blob/v2.8.4/lib/base/scriptutils.cpp#L232

[3]
https://www.icinga.com/docs/icinga2/latest/doc/17-language-reference/#boolean-values