Re-introduce lambdas for "set_if"
Commit d9f280e fixed a bug due to which "openshift_node_resources"
arguments with numeric values would be silently skipped. As part of that
change the templates for passing selectors to commands,
"openshift-arg-selector" and "openshift-arg-selector-py", were also
changed to use "set_if = "$…$"" rather than calling "len(macro(…))".

Unfortunately that change brought about another issue: "set_if" requires
its value to be either the string "true" or "false", or a value that can
be converted to an integral number. An argument is then only used when
the number is not zero. In the case of "openshift_node_resources" that
was the case as the variables are numbers.

Selectors are usually of the form "label=value". Checks using the
selector templates would have the selector silently dropped. Icinga logs
a warning ("Can't convert '…' to a floating point number") which was
missed during testing.

With this change we're changing all uses of "set_if" to use a lambda
function expanding the macro and passing the expanded value through the
"bool" function, thus implementing the same logic as the implicit
conversions to boolean[1].

[1]
https://www.icinga.com/docs/icinga2/latest/doc/17-language-reference/#boolean-values