new-app-and-wait: Remove only projects with recognized prefix
Ever since its inception the "new-app-and-wait" program would
automatically remove projects under the following conditions:

- Immediately if timestamp in project name can't be parsed

- After two times the maximum time allowed for application to become
  available when project name recognized prefix (usually "e2e-") and
  parseable timestamp (required to determine age)

- After 24 hours when project name prefix isn't recognized

It shouldn't happen, but if the end-to-end test code saw unrelated
projects due to faulty permissions on a cluster, it would also (attempt
to) remove those.

With this change the code only removes projects when it recognizes the
prefix. For all other visible projects merely a warning is emitted.