Consider all log heartbeat hits
Commit 0d11961 added a script to regularily check Elasticsearch in
OpenShift's logging component for a recent message sent by a cron job.
Due to the nature of Elasticsearch's match mechanism false positives may
be returned from the query (another message with the right keywords
would also match). That possibility is considered in the query by
returning multiple possible hits, but the evaluation code only used the
first hit. With this change all returned hits are considered.