Relax check for pod-to-node allocation
Ever since its initial implementation the "pod_node_alloc" plugin would
report a critical state if two or more pods with a given selectors were
allocated to the same node. With larger deployments, i.e. 5 pods, that
is usually acceptable. Even with deployments as small as three replicas
two instances on a single node are acceptable.

With this change the decision logic is changed to only complain if
either >50% or >3 of pods are on a single node. This way three-replica
deployments are considered okay if they have two pods on a single node.