object_stats: Output statistics of most recent cron job run
Commit 64c81d7 added a new plugin to collect statistics over objects in
a cluster. One of the object types is "CronJob" and the accompanying
"Job". In addition to success/failure per job it's also useful to have
those numbers per cron job (one of which may result in multiple jobs).
The statistics from the most recent successful job execution are output
along with the cron job itself.