object_stats: Support per-metrics selection of value
The original implementation of "check_openshift_object_stats" would
always add the value of metrics with the same name (e.g. emitting "foo"
twice with the values 12 and 45 would result in a metric "foo" with the
value 57).

With this change it becomes possible for metrics to define a selector
whereby an appropriate value is selected. "max" and "min" are supported.
These selectors will be used to output metrics from the most recent
successful cron job execution.