es_stats: Use Elasticsearch instance name in metrics
Since at least early 2018 OpenShift has used the deployment
configuration's name with its random suffix as the name for each
Elasticsearch instance. At the time of commit c04a18c that wasn't the
case yet. Unlike the "host" key, an IP address, the deployment
configuration name can easily be identified in a list of objects.

At the same time the functionality to strip prefixes from names is
amended with the ability to specify multiple and the default list is
extended to include "logging-es-data-master-".