es_stats: Compute global statistics
The "check_openshift_es_stats" would only generate per-node metrics and
apply limits to those. With this change the computation of the per-node
limits is unified with shared code and that is then used as a basis to
also compute cluster-wide statistics, allowing for limits to be applied
to alert on, for example, "more than 80% of disk space across all nodes
is used".