object_stats: Emit deletion timestamps as metric
Pods stuck in a terminal state would usually only be noticed by
a person. With this change a new metric is emitted when any of the
objects encountered by "check_openshift_object_stats" has a deletion
timestamp, defined by the API as "time when object is no longer expected
to exist", that is after the grace period.