new-app-and-wait: Add workaround token generation delayed
The creation of the builder service account token is delayed
on HA master environments. This causes rarely the situation
a build is started without an accessible access token and
stick in the state new. To reduce the chance to get into
this situation, a sleep of 3 seconds is added between
creating the project and starting the build.

For more information about this known bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1719792