obj_stats: Pods are either unscheduled or pending
Version 0.18.7 (commit d3013d1) started to emit metrics on how long
a pending pod has existed. An additional metric was generated for
unscheduled pods. These would usually show up alongside a "pending for
too long" alert which is redundant. With this version a pod is either
"pending" or "unscheduled", but not both.