obj_stats: Add option to always output named metrics
By default the "check_openshift_object_stats" program only outputs
metrics with defined limits. The primary reason is to reduce the amount
of metrics stored via Icinga, but also to reduce noise. With this change
two new options, "-a" and "-A", are introduced to always output metrics
with the given plain name or a name matching a regular expression.