obj_stats: Compute duration until pod is running
In commit bf6f22d an attempt was made to emit metrics to determine when
a pod was ready. As it turned out conditions are insufficient and one
must take container statuses into consideration.

With this change a new "running_after" metric on pods calculated for
pods which are running and which have no restarted containers. The
latter is because restarted containers, e.g. due to a failing health
check, have their "startedAt" timestamp reset.